requirements.txt
pyserial
opencv-python
numpy
Pillow
tensorflow==2.15.0
keras==2.15.0
scikit-learn
matplotlib
pandas
onnxruntime 